feat(core): move base extractor to new core package
