+++ src/core/__init__.py
@@ -1,0 +1,5 @@
+"""
+Core package for Coles scraper.
+"""
+
+from .base_extractor import HtmlExtractor
