+++ src/core/base_extractor.py
